+++ src/main/java/net/lopymine/mtd/gui/widget/list/AbstractSearchListWidget.java
@@ -17,7 +17,7 @@ public abstract class AbstractSearchListWidget<E extends Entry<E>> extends Abstr
 
 	public static final MutableText NOTHING_FOUND_TEXT = MyTotemDoll.text("text.nothing_found");
 
-	private final List<E> searchWidgets = new ArrayList<>();
+	private final List<E> savedWidgets = new ArrayList<>();
 	private boolean searching = false;
 
 	public AbstractSearchListWidget(int x, int y, int width, int height, int itemHeight) {
@@ -52,19 +52,14 @@ public abstract class AbstractSearchListWidget<E extends Entry<E>> extends Abstr
 	}
 
 	protected List<E> getWidgets() {
-		return this.searching ? this.searchWidgets : this.children();
+		return this.children();
 	}
 
-	//?} else {
-	/*@Override
-		protected E getEntry(int index) {
-		return this.searching ? this.searchWidgets.get(index) : super.getEntry(index);
-	}
-	*///?}
+	//?}
 
 	@Override
 	protected void drawMenuListBackground(DrawContext context) {
-		if (this.searching && this.searchWidgets.isEmpty()) {
+		if (this.searching && this.children.isEmpty()) {
 			int a = (this.getWidth() - this.getRowWidth()) / 2;
 			DrawUtils.drawText(context, NOTHING_FOUND_TEXT, this.getX() + a, this.getY(), this.getWidth() - a, this.getHeight() + 4);
 		}
@@ -72,26 +67,47 @@ public abstract class AbstractSearchListWidget<E extends Entry<E>> extends Abstr
 
 	@Override
 	public int getEntryCount() {
-		return this.searching ? this.searchWidgets.size() : super.getEntryCount();
+		return super.getEntryCount();
 	}
 
 	public void search(String string) {
 		this.setListScrollAmount(0);
 
 		if (string.isEmpty()) {
+			if (!this.savedWidgets.isEmpty()) {
+				this.setSelected(null);
+				this.children.clear();
+				this.children.addAll(this.savedWidgets);
+			}
 			this.searching = false;
+			this.updateCurrentWidgets();
 			return;
 		}
 
-		this.searchWidgets.clear();
-		for (E child : this.children()) {
-			if (searched(string, child)) {
-				this.searchWidgets.add(child);
+		if (!this.searching) {
+			this.savedWidgets.clear();
+			this.savedWidgets.addAll(this.children);
+		}
+		this.setSelected(null);
+		this.children.clear();
+
+		for (E child : this.savedWidgets) {
+			if (this.searched(string, child)) {
+				this.children.add(child);
 			}
 		}
 
-		this.searchWidgets.sort(getComparator());
+		this.children.sort(this.getComparator());
 		this.searching = true;
+		this.updateCurrentWidgets();
+	}
+
+	private void updateCurrentWidgets() {
+		//? if >=1.21.4 {
+		this.setScrollY(0);
+		//?} else {
+		/*this.setScrollAmount(0);
+		*///?}
 	}
 
 	protected abstract @NotNull Comparator<E> getComparator();